+++ src/main/java/video/api/java/sdk/domain/caption/CaptionClient.java
@@ -5,17 +5,18 @@ import video.api.java.sdk.domain.exception.ResponseException;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
+import java.util.Locale;
 
 public interface CaptionClient {
 
-    Caption get(String videoId, String lang) throws ResponseException;
+    Caption get(String videoId, Locale language) throws ResponseException;
 
-    List<Caption> list(String videoId) throws ResponseException;
+    Iterable<Caption> list(String videoId) throws ResponseException;
 
-    Caption upload(String videoId, File file, String language) throws ResponseException, IOException;
+    Caption upload(String videoId, File file, Locale language) throws ResponseException, IOException;
 
     Caption update(String videoId, CaptionInput captionInput) throws ResponseException;
 
-    void delete(String videoId, String language) throws ResponseException;
+    void delete(String videoId, Locale language) throws ResponseException;
 
 }